Refactor schemas for chat
