+++ backend/app/api/schemas/chat.py
@@ -1,26 +1,16 @@
-from datetime import datetime
-from typing import List, Union
+from typing import List
 from uuid import UUID
 
 from pydantic import BaseModel
 
-
-class MessageRequestSchema(BaseModel):
-    related_id: UUID
-    text: str
-
-
-class MessageResponseSchema(BaseModel):
-    custom_id: UUID
-    author_id: UUID
-    created_at: datetime
-    text: str
-    own: Union[bool, None]
+from .messages import MessageResponseSchema
+from .note import NoteResponseSchema
 
 
 class ChatResponseSchema(BaseModel):
     chat_name: str
     messages: List[MessageResponseSchema]
+    notes: List[NoteResponseSchema]
 
 
 class ChatsResponseSchema(BaseModel):
